Till here all method inside the video controller completed such as the updatevideodetail updatethumbnail deletevideo getvideobyId get allvideo with aggregation pipeline and sorting filtering pagination done...

+++ src/controllers/video.controller.js
@@ -1,4 +1,8 @@
 import {Video} from "../models/video.model.js"
+import { Comment } from "../models/comment.model.js"
+import { Like } from "../models/like.model.js"
+import { Playlist } from "../models/playlist.model.js"
+import { User } from "../models/user.model.js"
 import { ApiResponse } from "../helper/ApiResponse.js"
 import { ApiError } from "../helper/ApiError.js"
 import { asyncHandler } from "../helper/asyncHandler.js"
@@ -6,6 +10,7 @@ import {
     uploadOnCloudinary, 
     deleteOnCloudinary
 } from "../helper/cloudinary.js"
+import mongoose from "mongoose"
 
 
 
@@ -233,19 +238,109 @@ const deleteVideo = asyncHandler( async (req,res) => {
         throw new ApiError(400, "Please provide videoId");
     }
 
+    
+    
+    
     /*
     more work over here delete the video from the cloudinary and
     delete the watchHistory of User
     delete all the comment of this video
     delete all the likes of this video
     delete the video from the playlist
-
+    
     */
-
-    const video = await Video.findByIdAndDelete(videoId);
-
-    if (!video) {
+   
+   const video = await Video.findById(videoId);
+   
+   if (!video) {
         return next(new ApiError(404, "Video not found"));
+    }
+
+    if(video?.owner.toString() !== req.user._id.toString()){
+        throw new ApiError(400, "You are not authorized to delete this video");
+    }
+
+    try {
+        // deleting the videoFile and thumbnail from the cloudinary before deleting the video
+
+        const videoFilePublicId = video?.videoFile?.split("/").pop().split(".")[0];
+        const thumbnailPublicId = video?.thumbnail?.split("/").pop().split(".")[0];
+
+        const videoFileDeletefromCloud = await deleteOnCloudinary(videoFilePublicId);
+        
+        const thumbnailDeletefromCloud = await deleteOnCloudinary(thumbnailPublicId);
+        
+        console.log("Video file and thumbnail deleted from the cloudinary...",videoFileDeletefromCloud,thumbnailDeletefromCloud);
+    } catch (error) {
+        console.log("Error in deleting video from cloudinary", error);
+        throw new ApiError(500, "Internal server error");
+    }
+
+    // deleting the video from the Video entity
+    await Video.findByIdAndDelete(videoId);
+
+    // deleting the video comment from the Comment entity
+    await Comment.deleteMany(
+        { 
+            video: videoId,
+            owner: req.user._id
+        }
+    );
+
+    // deleting the video likes from the likes entity
+    await Like.deleteMany(
+        { 
+            video: videoId,
+            likedBy: req.user._id 
+            
+        }
+    );
+
+    // deleting the video from the playlist entity
+    await Playlist.deleteMany(
+        {
+            videos: videoId,
+            owner : req.user?._id,
+            
+        }
+    );
+
+    return res
+        .status(200)
+        .json(
+            new ApiResponse(
+                200,
+                {
+                    "Video Successfully deleted": video
+                },
+                "Video deleted successfully"
+            )
+        )
+
+
+})
+
+// togglePublishStatus controller...
+
+const togglePublishStatus = asyncHandler(async (req, res) => {
+    
+    const { videoId } = req.params
+
+
+    if(!videoId){
+        throw new ApiError(400, "Please provide videoId");
+    }
+
+    const video = await Video.findByIdAndUpdate(videoId,
+        {
+            $set : {
+                isPublished : !video.isPublished
+            }
+        }
+    )
+
+    if(!video){
+        throw new ApiError(400, "Video not found");
     }
 
     return res
@@ -256,22 +351,326 @@ const deleteVideo = asyncHandler( async (req,res) => {
                 {
                     video
                 },
-                "Video deleted successfully"
+                "Video updated successfully"
             )
         )
 
+})
+
+// Here we Get the all details about the video...
+
+const getVideoById = asyncHandler(async (req, res) => {
+    
+    const { videoId } = req.params
+    
+    if(!videoId){
+        throw new ApiError(400, "Please provide videoId");
+    }
+
+    const video = await Video.findById(videoId);
+
+    if(!video){
+        throw new ApiError(400, "Videos with this id not found");
+    }
+
+    const getVideoById = await Video.aggregate(
+        [
+            {
+                $match : {
+                    _id : new mongoose.Types.ObjectId(videoId)
+                }
+            },
+            // get owner details
+            {
+                $lookup : {
+                    from : "users",
+                    localField : "owner",
+                    foreignField : "_id",
+                    as : "owner",
+                    pipeline : [
+                        {
+                            $project : {
+                                username : 1,
+                                fullName : 1,
+                                avatar : 1,
+                                
+                            }
+                        },
+                        
+                    ]
+                }
+            },
+            
+            // get the likes of the videos
+            {
+                $lookup : {
+                    from : "likes",
+                    localField : "_id",
+                    foreignField : "video",
+                    as : "likes",
+                }
+            },
+            // get the comments of the videos
+            {
+                $lookup : {
+                    from : "comments",
+                    localField : "_id",
+                    foreignField : "video",
+                    as : "comments",
+                    pipeline : [
+                        {
+                            $project : {
+                                content : 1,
+                                owner : 1,
+                                createdAt : 1,
+                            }
+                        }
+                    ]
+                }
+            },
+            {
+                $addFields : {
+                    owner : {
+                        $first : "$owner"
+                    },
+                    
+                    likesCount : {
+                        $size : "$likes"
+                    },
+                    commentsCount : {
+                        $size : "$comments"
+                    },
+                    isLiked : {
+                        $cond: [ // here below same as ternary operator works...
+                            { 
+                                $in: [req.user?._id, "$likes.likedBy"] 
+                            },
+                            true,
+                            false
+                        ]
+                    }
+                }
+            },
+            {
+                $project: {
+                    title: 1,
+                    description: 1,
+                    videoFile: 1,
+                    thumbnail: 1,
+                    duration: 1,
+                    view: 1,
+                    isPublished: 1,
+                    createdAt: 1,
+                    updatedAt: 1,
+                    // owner: { 
+                    //     username: 1,
+                    //     fullName: 1,
+                    //     avatar: 1,
+                    //     email : 1,
+                    // }, 
+                    likesCount: 1,
+                    commentsCount: 1,
+                    owner: 1,
+                    comments: 1,
+                    isLiked: 1
+                }
+            }
+        ]
+    )
+
+    if(getVideoById.length === 0){
+        throw new ApiError(400, "Video not found");
+    }
+
+    // console.log("getVideoById", getVideoById);
+    console.log("Here is a getVideoById: " + JSON.stringify(getVideoById, null, 2));
+
+
+    return res
+        .status(200)
+        .json(
+            new ApiResponse(
+                200,
+                {
+                    getVideoById
+                },
+                "Video found successfully"
+            )
+        )
 
 })
 
 
 
+const getAllVideos = asyncHandler(async (req, res) => {
+    const { page = 1, limit = 10, query, sortBy, sortType, userId } = req.query
+    //TODO: get all videos based on query, sort, pagination
+
+    const pageNumber = parseInt(page)
+    const limitNumber = parseInt(limit)
+
+
+    const filter = {};
+    if(query){
+        const regex = new RegExp(query, "i"); // i stands for case insensitive
+        filter.$or = [
+            { title: regex },
+            { description: regex },
+        ]; 
+    }
+    if(userId){
+        filter.owner = userId
+    }
+
+    // const sortType = "asc"  // 1 means asc, -1 means desc
+    // const sortBy = "createdAt"
+
+    const sortOptions = {};
+    if(sortBy){
+        sortOptions[sortBy] = sortType === "asc" ? 1 : -1;  // sortOptions[sortBy] setting up the dynamic value like sortOptions : { createdAt : 1 or -1 }
+    }
+    console.log("sortOptions", sortOptions);
+
+    // here Video Aggregate pipeline
+    const videos = await Video.aggregate(
+        [
+            {
+                $match : filter
+            },
+            {
+                $sort: sortOptions // Sort by the specified field
+            },
+            {
+                $skip: (pageNumber - 1) * limitNumber // Skip records for pagination
+            },
+            {
+                $limit: limitNumber // Limit the number of records returned
+            },
+
+            // getting the owner of that video
+            {
+                $lookup : {
+                    from : "users",
+                    localField : "owner",
+                    foreignField : "_id",
+                    as : "owner",
+                    pipeline : [
+                        {
+                            $project : {
+                                fullName : 1,
+                                username : 1,
+                                avatar : 1
+                            }
+                        },
+                        
+                    ]
+                }
+            },
+            // getting the likes of that video
+            {
+                $lookup : {
+                    from : "likes",
+                    localField : "_id",
+                    foreignField : "video",
+                    as : "likes",
+                }
+            },
+            // getting the comment of that video
+            {
+                $lookup : {
+                    from : "comments",
+                    localField : "_id",
+                    foreignField : "video",
+                    as : "comments",
+                    pipeline : [
+                        {
+                            $project : {
+                                content : 1,
+                                createdAt : 1,
+                                owner : 1,
+
+                            }
+                        },
+                        
+
+                    ]
+                }
+            },
+            {
+                $addFields: {
+                    likesCount : {
+                        $size : "$likes"
+                    },
+                    commentCount : {
+                        $size : "$comments"
+                    },
+                    // comment : {
+                    //     $first : "$comments"
+                    // }
+                }
+            },
+            {
+                $project: {
+                    
+                    title: 1,
+                    description :1,
+                    videoFile :1,
+                    thumbnail : 1,
+                    owner : 1,
+                    duration : 1,
+                    views : 1,
+                    isPublised : 1,
+
+                    comments : 1,
+                    likesCount : 1,
+                    commentCount: 1,
+
+                    createdAt : 1,
+                    updatedAt : 1,
+
+
+                }
+            }
+
+        ]
+    )
+
+    if(videos.length === 0){
+        throw new ApiError(400, "Videos not found");
+    }
+
+
+    const totalVideos = await Video.countDocuments(filter);
+    const totalPages = Math.ceil(totalVideos / limitNumber);
+    console.log(10/5);
+
+    return res
+        .status(200)
+        .json(
+            new ApiResponse(
+                200,
+                {
+                    videos,
+                    totalVideos,
+                    totalPages,
+                    currentPage: pageNumber,
+                    limit: limitNumber,
+                },
+                "Videos found successfully"
+            )
+        )   
+})
 
 export {
     videoHealthCheck,
     createVideo,
     updateVideoDetails,
     updateVideoThumbnail,
-    deleteVideo
+    deleteVideo,
+    togglePublishStatus,
+    getVideoById,
+    getAllVideos,
+
 }
 
 
@@ -282,39 +681,7 @@ const getAllVideos = asyncHandler(async (req, res) => {
     //TODO: get all videos based on query, sort, pagination
 })
 
-const publishAVideo = asyncHandler(async (req, res) => {
-    const { title, description} = req.body
-    // TODO: get video, upload to cloudinary, create video
-})
 
-const getVideoById = asyncHandler(async (req, res) => {
-    const { videoId } = req.params
-    //TODO: get video by id
-})
-
-const updateVideo = asyncHandler(async (req, res) => {
-    const { videoId } = req.params
-    //TODO: update video details like title, description, thumbnail
-
-})
-
-const deleteVideo = asyncHandler(async (req, res) => {
-    const { videoId } = req.params
-    //TODO: delete video
-})
-
-const togglePublishStatus = asyncHandler(async (req, res) => {
-    const { videoId } = req.params
-})
-
-export {
-    getAllVideos,
-    publishAVideo,
-    getVideoById,
-    updateVideo,
-    deleteVideo,
-    togglePublishStatus
-}
 
 */
